requirements.txt
beautifulsoup4==4.11.1
pandas==1.3.5
pyfiglet==0.8.post1
pyperclip==1.8.2
pytest==7.1.2
python-dotenv==0.21.0
requests==2.26.0
termcolor==1.1.0
